src/requirements.txt
aniso8601==4.0.1
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
dlib==19.16.0
Flask==1.0.2
Flask-Cors==3.0.7
flask-restplus==0.12.1
Flask-Script==2.0.6
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.0
more-itertools==5.0.0
numpy==1.15.4
opencv-python==3.4.5.20
Pillow==5.4.0
pluggy==0.8.0
py==1.7.0
pytest==4.0.2
pytz==2018.7
six==1.12.0
Werkzeug==0.14.1
